Fix for obviously spurious points-to results
The pointed-to allocation sites are now tracked inside PointsTo terms.
This differs a little bit from the usual presentation, but it does not
impact correctness. It makes it unnecessary to keep track of all
constructed (PointsTo) terms to recover the points-to relation.

The spurious points-to results were from a misinterpretation of the term
equality axiom. It says that if two terms, ↑α1 and ↑α2, are found to be
equal, then α1 and α2 must also be equal.
In a sense I had implemented this axiom as a bi-implication instead of
just the implication.